+++ lib/data/remote/service/connectivity_service.dart
@@ -1,9 +1,10 @@
 import 'dart:async';
 
 import 'package:application_base/core/service/logger_service.dart';
-import 'package:application_base/core/service/service_locator.dart';
 import 'package:application_base/domain/subject/network_subject.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
+import 'package:injectable/injectable.dart';
+import 'package:meta/meta.dart';
 
 /// Connectivity changes are no longer communicated to Android apps
 /// in the background starting with Android O (8.0).
@@ -12,18 +13,14 @@ import 'package:connectivity_plus/connectivity_plus.dart';
 ///
 /// On iOS simulators, the connectivity types stream might not update
 /// when Wi-Fi status changes.
+@lazySingleton
 final class ConnectivityService {
   ///
-  ConnectivityService._();
+  @visibleForTesting
+  ConnectivityService(this._connectionSubject);
 
   ///
-  factory ConnectivityService.singleton() => _instance;
-
-  ///
-  static final _instance = ConnectivityService._();
-
-  ///
-  final _connectionSubject = getIt<NetworkSubject>();
+  final NetworkSubject _connectionSubject;
 
   ///
   final List<ConnectivityResult> _connectivityList = [];
@@ -54,6 +51,7 @@ final class ConnectivityService {
   }
 
   ///
+  @disposeMethod
   void dispose() {
     _subscription?.cancel();