+++ lib/domain/subject/network_subject.dart
@@ -1,25 +1,24 @@
 import 'dart:async';
 
 import 'package:application_base/data/remote/const/network_event.dart';
+import 'package:injectable/injectable.dart';
+import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 
 export 'package:application_base/data/remote/const/network_event.dart';
 
 ///
+@lazySingleton
 final class NetworkSubject {
   ///
-  NetworkSubject._();
-
-  ///
-  factory NetworkSubject.singleton() => _instance;
-
-  ///
-  static final _instance = NetworkSubject._();
+  @visibleForTesting
+  NetworkSubject();
 
   ///
   final _networkSubject = PublishSubject<NetworkEvent>();
 
   ///
+  @disposeMethod
   void dispose() {
     _networkSubject.close();
   }