+++ lib/presentation/view_model/access_vm.dart
@@ -1,16 +1,13 @@
 import 'package:application_base/core/service/logger_service.dart';
 import 'package:flutter/foundation.dart';
+import 'package:injectable/injectable.dart';
 
 ///
+@lazySingleton
 final class AccessVM with ChangeNotifier {
   ///
-  AccessVM._();
-
-  ///
-  factory AccessVM.singleton() => _instance;
-
-  ///
-  static final _instance = AccessVM._();
+  @visibleForTesting
+  AccessVM();
 
   ///
   bool _isGranted = false;